requirements.txt
PyYAML
numpy
torch
pytest
psutil
websocket-server
websocket-client
